+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Fixes
 - `from_regex/1`:
     - Ignore inline comments `(?# )`
+    - Add support for `:extended` (`/x`) modifier
 
 ## 0.2.0 [2026-03-30]